[virtualization][vsock] Move socket outside of RefCell
Async read and writes require a mutable reference, but we were never
using the async functionality due to the credit system. This means we
can just drop down the sync socket when reading and writing instead, and
do away with the RefCell entirely.

The key benefit here is that the RX path won't need a write lock on the
state, just a read lock.

Bug: 97355
Change-Id: I303632fc06a9ebdfecd855025b0c35072693cd50
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698224
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: David Hastings <dahastin@google.com>