[ffx][profile mem] Add flag that outputs the raw JSON written by memory monitor.
The purpose of this command is to facilitate investigating parsing error
that the command `fx profile memory` encounters. It does so by printing
the output of Memory Monitor, see https://fuchsia.googlesource.com/fuchsia/+/refs/heads/main/src/developer/memory/monitor/monitor.cc#278

Because this command is to facilitate future debugging, no guarantee is
made on the stability of format of the output.

Change-Id: Idd5027a886d0be624305ef5ad2a71689881f84c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697904
Reviewed-by: Guillaume Chelfi <chelfi@google.com>
Commit-Queue: Jean-François Geyelin <jif@google.com>