[debugger] Convert to using DwarfUnit.
Attempting to use the UnitSymbolFactory for the first time revealed that
the offset we actually want is from the DwarfUnit, not the CompileUnit.

This patch adds a way to get the DwarfUnit from the CompileUnit, removes
the now-unnecessary DIE offset on CompileUnit, and adds clarifying
comments around the difference between DwarfUnit and CompileUnit.

Bug: 79529

Change-Id: I11043429a0f53330660f40a9063a35900efa346b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698566
Commit-Queue: Brett Wilson <brettw@google.com>
Reviewed-by: Dangyi Liu <dangyi@google.com>