[fdf][bind] Migrate fidl_offers in props to gened
Migrate the device properties that are added from the fidl offers on a
device to the new generated bind libraries. Previously the property had
a value of true, now it is the enum from the bind library we generate
from the FIDL protocols.

Bug: 100242
Tests: Updated the integration and unit tests and confirming they work
Change-Id: If5442bd31967fd4bcbbd990a7447c6c7a4f08a06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698624
Reviewed-by: Sarah Chan <spqchan@google.com>
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>