[component][fuzzing] Re-enable realmfuzzer integration tests
This CL re-enables the integration tests that flaked in coverage builds.
The underlying cause of the flake has been observed to be a flawed
assumption: based on informal conversations, it was assumed that an
explicit route (for `fuchsia.debugdata.Publisher`) would take
precedence over the implicit route made available the "test-env"
environment's debug field in coverage builds.

This is not always the case, and the integration tests were observed to
(frequently) connect to test_manager's debug_data component instead of
the fake fuzz_coverage component provided by the test. As a result, the
intentionally crashed process would not be monitored and the test would
fail.

Without a good way to force the selection of the explicit route, the
most effective solution is just to skip these tests for coverage builds.
Code coverage is less important for these integration tests than for the
corresponding unit tests. Functionality is still tested by every
non-coverage build.

Bug: 103781
Multiply: fuchsia-pkg://fuchsia.com/component-fuzzer-integration-tests#meta/fuzzer-integration-tests.cm
Change-Id: I67ca7ecdc0e672d8b2ef52da626687c8662a0337
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698252
Reviewed-by: Cameron Finucane <eep@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>