[virtualization] Implement RESOURCE_ATTACH_BACKING
The RESOURCE_ATTACH_BACKING command allows for the guest to specify a
set of guest-physical pages that are associated with a resource. These
guest-physical pages need not be physically contiguous so the backing
memory is provided in the form of a scatter-gather list.

Once the backing memory is associated with a resource, the driver can
then issue TRANSFER_TO_HOST_2D commands to copy data from the backing
memory into the host-private resource (ex: a vmo shared with Flatland).

Bug: 102870
Test: fx test virtio_gpu_unittests
Change-Id: I131e07b5c463e283e6fece6c5953f41073b2ada1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695690
Reviewed-by: David Hastings <dahastin@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>