[archivist][tests] Use the same archivist for testing v1 protocols
Instead of two separate archivist, we can use the same one to test
protocols that are only used in the v1 world.

One log is reduced to "warn" severity, as the "log connector" protocol
is typically not available (provided by appmgr) unless the test itself
is mocking it and providing it. Instead of allowing "ERROR" logs in the
test we can just decrease severity to warn.

Change-Id: If7a9c2c774821d4f1d9a4d39845f011667d8054e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698587
Reviewed-by: Christopher Johnson <crjohns@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>