[settings] Add fidl_storage to factory
Fidl storage factory support is added with this CL. The Fidl storage
factory does not differ much from the current DeviceStorageFactory used
for stash. It only instantiates a different backing storage and has
different serialization requirements for the types it can prepare for
storage.

The /data/storage directory is used to store the data files, and the
storage agent is now able to choose between stash and persistent fidl
storage when writing. The agent keeps track of both factories, and will do
so until all settings have migrated away from stash.

The vast majority of file changes are are updating the storage trait
impls for the controllers to mark which storage is being used to track a
particular setting.

Currently, no settings have been converted to using fidl storage. This
is mostly plumbing in preparation for the first setting to be converted
to persistent fidl.

Bug: 91407
Test: fx test -o setui_service_tests
Change-Id: I53897687759ba4ec97cc335e2bed6980ffce1cfe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693310
Reviewed-by: William Xiao <wxyz@google.com>
Commit-Queue: Paul Faria <paulfaria@google.com>