[securemem] Adjust securemem zeroing
We need to zero out buffers via the TEE that come from the REE. When
calling into the TEE we select a max zeroing length to ensure that we
don't spend too long in the TEE and could cause problems with
scheduling. Increasing the zeroing length reduces the overhead of
calling into the TEE. 1 MiB zeroing length has shown to not cause
scheduling problems and also cut the zeroing time by roughly 1/2.

Bug: 103225
Bug: b/235629262
Change-Id: I67c236729d8bf5e806d9f56e5c8c02ab1fa8cbbd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/696709
Commit-Queue: Stefan Bossbaly <stefanbossbaly@google.com>
Reviewed-by: Dustin Green <dustingreen@google.com>