[hypervisor] Refactor FPU save and restore
Move FPU save and restore out of the core VCPU loop, so that we can
speed up VM entry and exit.

We only need to save and restore the FPU state when we are returning
back to host user-space, and when migrating CPUs, this is automatically
handled by the context switch. In all other cases, we can avoid the cost
of the FPU save and restore.

Bug: 95763
Change-Id: I7cf9482d6e59085451e000eb9edf5527e47aad7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/697823
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Fuchsia-Auto-Submit: Abdulla Kamar <abdulla@google.com>
Commit-Queue: Abdulla Kamar <abdulla@google.com>