[overnet] rewrite stream framer library
This rewrites the stream-framer library from being a complex state machine
into a synchronous helpers that manage serializing and deserializing the
bytes. In addition, it lifts management of the deframing timeout from the
Deframer and over to the single client that uses this functionality. This
allows further simplification of the Deframer.

Change-Id: Ic5d1bcda85803c4cedc42317b2687e099eb04a83
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699129
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>