[starnix] Reclaim shared memory in binder
This change adds logic to reclaim shared memory when it is freed, and
also updates the allocator to use the reclaimed memory.

The new allocator works as follows:

  1. Check if the allocation fits between the current offset and either
     the end of the shared vmo, or the start of the next allocation.
  2. If the allocation does not fit, attempt to "wrap" the allocator
     back to the start of the vmo. If the allocation does not fit at the
     start of the vmo, then return ENOMEM. If it does, allocate at the
     start and update the offset accordingly.

Clients will retry on ENOMEM, and given the pattern of allocations this
allocator should be sufficient for quite some time (since the
SharedMemory instances are shortlived).
.

Change-Id: I5e85c151cba0fa780b7502e622e60da4ba25acff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698788
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Commit-Queue: Kevin Lindkvist <lindkvist@google.com>