[storage][perf] New filesystem benchmark framework
This framework is intended to replace the existing
start-storage-benchmarks + odu + tracing combo.

The main reason for the new framework is the ability for the benchmarks
to remount the filesystem. Remounting the filesystem should clear any
cached data from the setup phase of the benchmark so we can benchmark
cold reads and writes.

Bug: 94248
Change-Id: I754027e11dbf62c72077146611422edb86939a92
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695563
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>