[fidl][server_suite] GIDL-like DSL
This CL introduces a GIDL-like DSL for defining byte arrays and a
mechanism to write and read these byte arrays from a channel.

This is motivated by a desire to introduce more complex tests to the
test suite. In this CL, two tests were added that cover basic usages of
error syntax results.

Bug: 102339
Change-Id: I0779ce7d55a1972ba674d57acd6f2265702ecddd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698429
Reviewed-by: Yifei Teng <yifeit@google.com>
Commit-Queue: Benjamin Prosnitz <bprosnitz@google.com>