[build] Don't always build the full SDK
We should build the SDK in SDK builders for instance.
Most often, we build a product and we use some SDK tools for build
actions, but don't need to build the full SDK.

This is important especially because the SDK cross-compiles host tools
with all supported host toolchains, rather than with just the toolchain
for the build host. For instance if building and testing core.x64 there
is no need to build ffx for host_arm64.

This makes some clean builds >20% faster!!
This also significantly cuts down on iteration times when working on ffx
or ffx plugins or their dependencies.

Bug: 101211

Change-Id: I5e14222e193c1302e2a3660ea5587f50a3a32495
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698968
Reviewed-by: Nathan Mulcahey <nmulcahey@google.com>
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Commit-Queue: Shai Barack <shayba@google.com>
Reviewed-by: Oliver Newman <olivernewman@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Reviewed-by: Megan Batty <mgnb@google.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>