[ffx][config] Environments know their own path.
This makes it so that an Environment knows where it came from, so you
don't have to keep track of that separately. From this starting point
we'll be able to build up some additional capabilities in the
environment object as well.

Change-Id: If99fecc9371420d242d47690041f426408a850a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699075
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>