[ffx][config] Move query parsing out of storage
This moves the query parsing logic completely out of the storage::Config
object, making it a great deal simpler.

As part of that, it moves a bunch of things that were lying all over the
codebase and puts them in more logical places (like get_config from the
api root to query and validate_write_query from Config to query). We get
to only validate the queries in one place as well, which is nice.

Change-Id: Id15fd345f8ba9995ab3328f5977a9c1688eea05a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699076
Fuchsia-Auto-Submit: Megan Batty <mgnb@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>