[stream][test] Reduce iterations in ReadShrinkRace
Reduce the number of iterations and wait time in ReadShrinkRace, since
it is an inherently slow test.

https://fxrev.dev/698442 initially bumped the number of iterations of a
couple of tests to better test some races. This test is causing some
other builders to time out on tests because of the number of iterations
it is now doing.

Change-Id: I550d4c8ee964a7375c887807d020d4be5fd07b33
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699366
Reviewed-by: Drew Fisher <zarvox@google.com>
Fuchsia-Auto-Submit: Chris Tam <godtamit@google.com>
Reviewed-by: Marc-Antoine Ruel <maruel@google.com>