[cts][pm] Support specifying archive name
Add support for specifying archive name in pm_archive (and
fuchsia_packge_archive) instead of always using the package name. This
is helpful in the CTS because cts_fuchsia_test_package wants to create a
normal fuchsia_test_package and a fuchsia_package_archive and put them in a
group(target_name). This causes the archive to have the intermediary
target's (the fuchsia_test_package) name, which forces users to include
another build target in dependencies.

Bug: 104198

Change-Id: I597333d8cd9de04950e58dd128590fdac60090fc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699325
Commit-Queue: John Shamoon <johnshamoon@google.com>
Reviewed-by: Erick Tryzelaar <etryzelaar@google.com>