[kernel][arm64] arm64 usercopy variant
This introduced a slight rewrite of usercopy.
The primary goal is to enable switching to ldr/stt instructions.

This CL also introduces arm64_user_copy_test GN template, for
adding a test of a variant, based on the loop unrolling that was
originally done for memcpy.

The next steps involve allowing to align either src or dst on the
loop prologue.

The kernel implementation has not changed at this point.

Bug: 59284
Test: fx test -v -o arm64-user-copy-tests
Change-Id: Ib0de053d5cb2c26c9c5bf8bb83b875f5de95d89f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/693863
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Travis Geiselbrecht <travisg@google.com>