[fidl] Re-organize compatibility test directory
The goal here is to minimize the "differences" between the various
implementation. Obviously, there will still be some differences in
naming, build rules, file structure, and the code itself, but the ideal
is to get to a situation where all of the "incidental" differences
between any two binding language implementations are removed.

Some examples of this:
- Each implementation is moved into its own sub directory, with its own
  BUIDL.gn file, its own /meta directory for its manifest, and so on.
- Every binding binary is named 'main.[EXT]`
- Every server has an identical name format
- Etc...

Note: due to the shuffling of some names, the CMX allowlist must be
updated as well. The modified files will be removed in a later CL in
this chain, one which this refactor makes easier to land.

Test: fx test fidl_compatibility_test
Bug: 91343
Change-Id: Ia41c1b146abf879e24580e14a7e23131e9bc3ddf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698994
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Shai Barack <shayba@google.com>