[fidl] Migrate compatibility test runner to CFv2
In order to better accommodate CFv2, and to simplify the launching logic
of these tests a bit, the "runner <-> proxy <-> server" model has been
slightly modified by this change. Previously, the runner was responsible
for launching the proxy, which was in turn responsible for launching the
target server it was proxying to. This requires a large number of
capabilities to be passed around between the runner and its various
proxy implementation targets (there are currently 6 of these), so this
setup has been refactored.

Instead, the runner now "eagerly" launches BOTH the proxy and the
server.  The proxy is no longer responsible for launching anything, and
now simply relies on the service directory to properly locate the
already running target server.

This change in launch responsibility means that the in-protocol
`forward_to_server`-like fields are now redundant, since the proxy no
longer needs to know the server's URL to properly connect to it. These
fields can probably be removed in a future change (see:
https://fxbug.dev/104132).

Additionally, this change disables the Dart implementation of the
`fidl_compatibility_test`. This is because I couldn't figure out how to
pass the `dart_jit_runner` runner config from the converted test runner
`.cml` file to the old `dart-impl.cmx`. I've left a TODO, and the
subsequent CL in this chain that converts all of the impls to `.cml`
will re-enable this test as well.

Test: fx test fidl_compatibility_test
Bug: 104132
Bug: 91343
Change-Id: I6c5970c16beace253caeb69fab7aa59dff3368a5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698998
Reviewed-by: Ian McKellar <ianloic@google.com>
Commit-Queue: Alex Zaslavsky <azaslavsky@google.com>