[bt][hfp] Fix SLC Init three-way calling path
When the peer does not support three-way calling but does support HF
indicators, we were expecting an AT+CIND? instead of the correct
start to the HF indicators exchange.

Add tags to the logs for all parts of this component for ease of use.

Rename some of the SLC states to more accurately reflect the states we
are in, and add some comments from the spec for the triple-ending state
of SLC Initialization.

Bug: 74091
Test: updated tests, manual test with peer

Change-Id: I70cd74239e7fb7666d71ad5a189d79fcf56c2994
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/688452
Reviewed-by: Matthew Kehrt <kehrt@google.com>
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>