[virtualization] Allocate resources with sysmem
In order to import host resources into Flatland, we need to allocate
those buffers using sysmem.

This complicates our unit tests some since we can't access sysmem
without making our unittests non-hermetic. As a short term solution we
continue to use the non-sysmem in unit-tests. The main implication is we
won't have an associated import token for the VMO, but the unit tests
will not be interfacing with flatland or sysmem anyways.

Bug: 102870
Test: fx test virtio_gpu_unittests
Change-Id: I2e20f74f702b05c7e94b93a33614d1647ba161f0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698089
Reviewed-by: David Worsham <dworsham@google.com>
Reviewed-by: David Hastings <dahastin@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>