[ffx] Clean up stop message when nothing's running
The attached bug indicates that when we switched to the common
get_engine_by_name function for resolving emulator names, we missed a
case in "stop" when no name is provided. The error message from that
function was being discarded in favor of one about "deserialization".
While it's true the code couldn't deserialize an engine, this was not a
useful way to describe it to the user.

This CL adds a check for the case where no engine name was resolved, and
prints a separate message from the one when the engine is resolved but
fails to be deserialized.

Bug: 103534
Change-Id: I4ace3e9ee2807e9dc1e7cc56a6c786976e470042
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699854
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>