[ffx] Make log paths relative to PWD
Previously, if a relative path was passed to ffx emu for the log
parameter, it would be resolved by the emulator process, resulting in
the log file living in the emulator's working directory. This doesn't
fit with most developers expectations, as shown by the attached bug.

This CL explicitly sets the log file to be relative to the current
working directory of the FFX tool before launching the emulator, so it
will be in line with the user's expectations. Absolute paths are always
treated as absolute, so no changes are needed with the handling of those
paths.

Bug: 103652
Change-Id: I1af11902aff57c31837852f4b9baf440638c9778
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699843
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>