[docs][performance] Change "Catapult" references to refer to "Chromeperf"
In Fuchsia we had started off referring to Chromeperf as "the Catapult
performance dashboard", or just "Catapult", but the name "Chromeperf"
is now preferred.

This only changes the references in the "user guide" page and the
files that link to that.  It doesn't change other references to
"Catapult" yet.

Add a redirect for the old page.

Add an explanation of the original use of the "Catapult" name.

Change-Id: Iefd7dc6a7f3eb005e4c5bb61542287a804f737c5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698605
Reviewed-by: Vivian Tran <viviantran@google.com>
Reviewed-by: Nick Van der Auwermeulen <nickvander@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>
Reviewed-by: Fadi Meawad <fmeawad@google.com>