[utest][debugger] Strip tag from register before reading at that addr
`test_memory_ops` will fail under hwasan because it attempts to read
from a tagged value via a call to zx_process_read_memory, which should
not accept tags under the address tagging ABI. The address we read
from is actually tagged from a segfault on a tagged pointer in a
separate function. In this particular case, the tag can be ignored
and the test should still be able to read from that pointer.

Bug: 57112
Change-Id: Id5f354eb347929a04266727f65c57b3249660e2a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699043
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>