[bt][avrcp] Report correct play status after time
When sending back from a GetPlayStatus command, the target would send
back the last computed play status, which often was just at the
beginning of the track, or the last time play / pause was sent.  Update
the current playback position based on the current time when
GetPlayStatus is called.

Update PlaybackRate to be a newtype for TimelineFunction and update the
related code.

Bug: 102195
Test: updated tests, fx test bt-avrcp-target-tests

Change-Id: Id1f8ffae67f815b8bfa71f40b8e424e7afc146a8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699755
Reviewed-by: Zac Bowling <zbowling@google.com>
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>
Commit-Queue: Marie Janssen <jamuraa@google.com>