[userboot] Svc stub passed to processes
An additional handle with name /svc and of type PA_NS_DIR
is sent to launched processes, this allows forwarding the other end
of the svc handle to a component at a later stage.

This would enable instrumentation of core-tests and component manager
(as part of a boot sequence).

Added test that verifies individual aspects of the protocol, including
the publish debug data integrity done through sanitizer_publish_data.

Due to C Bindings not being supported for fuchsia.io, we had to manually
decode the message, the bug with one shot decoding in LLCPP(not
involving an async dispatcher and a custom server implementation) has
been added to the manual decoding, such that it easier to cleanup
afterwards.

Bug: 101831
Test: fx run-zbi-test early-boot-single-process-test
Change-Id: Ie98919ae0f43e8da3ccbf4af6d69a0c1fb1411ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/675469
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>