[virtualization-net-tests] Adding fix for virtio net test flake
A previous change fixed a target issue, but left an issue on the host.
It is possible for the host to send a packet to the target before the
target's netdevice's receive buffers are ready. Add a loop to retry in
case this occurs.

Change-Id: Ia0149b2146a7c27b4c3b703ca609ee60e2eda25c
Multiply: virtualization-net-tests
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698667
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Joon-Hyuk Han <hanjh@google.com>