[scenic] Handle display rotation in gfx scene manager
Rotation: We read a config file /config/data/display_rotation and: (1)
give it to the compositor, (2) set the correct rotation/translation for
the pointerinjector_view_holder. This should be the same logic as root
presenter:
https://cs.opensource.google/search?q=display_startup_rotation_adjustment_&sq=&ss=fuchsia%2Ffuchsia:src%2Fui%2Fbin%2Froot_presenter%2F

Scale: We now set scale on the pointerinjector_view_holder instead of
using scene.set_scale.  IIUC this is necessary so that e.g. a11y
highlights will be drawn an the correct scale.

Note: Credit dworsham@ for a lot of this: I split out the
rotation/scale-related changes from fxr/688687, fixed a few issues, and
added rotation to the pointerinjector-config-test.

Fixed: 44245
Fixed: 45541
Change-Id: I7c5331409b5353abcd090c87733a89f2d7667c35
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/695918
Reviewed-by: Alexander Brusher <abrusher@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Michael LoPiccolo <milopiccolo@google.com>