[sdk] Remove the tools from the legacy host tools directory
Currently some tools are uploaded to the top level directory in the
tools/ directory. For example:
* tools/symbol-index <---- wrong, getting deleted
* tools/arm64/symbol-index
* tools/x64/symbol-index

tools/x64/symbol-index and tools/symbol-index are the same and
all consumers are using the tools in either tools/x64/* or
tools/arm64/*.

This commit removes the tools from the top level directory and
instead partners will use the tools in either tools/x64 or tools/arm64.

Clone of fxr/661083

Bug: 42999
Change-Id: I26324c8446c8dbaecdbce7aad4c57547b2597c7e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/692966
Reviewed-by: Andrew Davies <awdavies@google.com>
Reviewed-by: Dave Schuyler <dschuyler@google.com>
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Sébastien Marchand <sebmarchand@google.com>
Commit-Queue: Omer Lev-ran <omerlevran@google.com>