[ui] Add use_scene_manager_as_scene_owner flag
This flag enables developers to permute the UI realm between
scene_manager and root_presenter on products as we attempt to migrate
all products to using scene_manager.

When the migration is complete and we delete root_presenter, we will
also delete this flag.

Fixed: 98781
Change-Id: I05dc49c0c33db98be1295aca94f7951be77fd87c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698977
Reviewed-by: Aaron Wood <aaronwood@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Commit-Queue: David Worsham <dworsham@google.com>
Reviewed-by: Sahil Madan <sahilmadan@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Alice Neels <neelsa@google.com>