[ffx] Allow overriding ffx's command name.
This change makes it so that ffx can read an environment variable for a
command name override, letting it show useful help when invoked as
`fx ffx` in tree or `tools/ffx` in the sdk, while allowing the normal
behaviour as before otherwise.

Also broke down the copy/pasted code from argh and made it a bit more
testable so I could verify this behaviour in tests.

Bug: 101039

Change-Id: I0901e1e4a1dd667912d53541933fbf2361380fe6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699982
Commit-Queue: Megan Batty <mgnb@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>