[benchmarks] Run dart-fidl-benchmarks using Components v2 (second try)
This uses the Dart test runner component that was not available when
the earlier version of this was landed
(https://fuchsia-review.googlesource.com/c/fuchsia/+/672283).

The Dart test runner had a problem with granting access to the
custom_artifacts directory (see https://fxbug.dev/102330).  The
problem seems to have been fixed with a recent roll of the Dart test
runner, but we're not sure what fixed the problem.

Update the README.md by removing instructions for running the test
that are no longer applicable, because they pre-date the plumbing in
fidl_microbenchmarks_test.dart for reporting the results through the
performance testing infrastructure, and I don't think it makes sense
for each test to have instructions on how to run it.  These
instructions were originally in the Topaz repo under
tests/dart_fidl_benchmarks/README.md.

Bug: 98667
Test: fx set terminal.x64 --release
  + fx test --e2e host_x64/fidl_microbenchmarks_test
Change-Id: Iee6615b7bcb508663186886134d827c3df27ea9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/687989
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Commit-Queue: Mark Seaborn <mseaborn@google.com>