[dhcp] Pass updated UpdatedAt to acquiredFunc
The previous refactor of DHCP client made it so that the updated
value of `Config.UpdatedAt` wasn't passed to `Client.acquiredFunc`
when it was called. This meant that the computation for the
valid-until address property for addresses assigned by the DHCP client
was wrong for all renewals (we were always adding the lease duration to
the time the lease was originally obtained, not the subsequent
renewal times).

Modified a unit test to act as a regression test for this.

Test: fx test --test-filter TestClientUpdateInfo netstack-dhcp-gotests
Change-Id: I05152e6279f4b348e6279e8cd000e756372b08eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700366
Reviewed-by: Nate Hurley <nahurley@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Tony Gong <gongt@google.com>