[ffx] Make config flag relative to CWD
Previously, if a relative path was passed to ffx emu for the config
template parameter, it would be resolved after the process had changed
the working directory, resulting in the inability to use files local to
the user's actual CWD. This doesn't make sense in most use cases, as
shown by the attached bug.

This CL explicitly sets the template files path to be relative to the
current working directory of the FFX tool before launching the emulator,
so it will always be in line with the user's expectations. Absolute
paths are always treated as absolute, so no changes are needed with the
handling of those paths.

Bug: 95945
Change-Id: I4faa2626443cf6802e32e6e07d3e0d9d5c474898
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699890
Reviewed-by: Clayton Wilkinson <wilkinsonclay@google.com>
Commit-Queue: Daniel Dougherty <doughertyda@google.com>