[archivist] Remove obsolete test
We already have coverage for this in the integration tests and it's
using the TestWithEnvironment fixture we want to remove.

Additionally the cpp-logs test is already using the structured backend
which is the default. So we don't need to have the same test twice.

Change-Id: I3b16cb7eb70bb801c9421b24f71ec506f8dbcda0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700786
Reviewed-by: Christopher Johnson <crjohns@google.com>
Commit-Queue: Miguel Flores <miguelfrde@google.com>
Fuchsia-Auto-Submit: Miguel Flores <miguelfrde@google.com>