[governance] Move area descriptions into YAML.
This old structure was kinda confusing, with the descriptions living
in README.md and the metadata living in _areas.yaml. When the
"Graphics" area got semi-accidentally renamed, README.md didn't get
renamed, and so the description simply didn't show up on the API
Council page.

The downside is that the descriptions have to use HTML, rather than
markdown. However, this is already how the glossary works, so that's
not too unusual.

Also adds the API primary/secondary to the description for each area.

Change-Id: Ib7742689dcd5f3ec151986ac5eafcafc39238988
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/698330
Reviewed-by: Alice Neels <neelsa@google.com>
Commit-Queue: Hunter Freyer <hjfreyer@google.com>