[dhcpv6] Abstract function for processing options
Add a process_options function which processes and validates options.
The function is passed the exchange type (valid combinations of
request and response message types), and therefore it is able to
perform validation specific to each message exchange, e.g. IA
options are not valid when receiving a Reply in response to an
Information-request message.

The Advertise message handler in ServerDiscovery state and the Reply
message handler in InformationRequesting state have been changed to
use the new function. The Reply message handler in Requesting state
was left untouched, as significant changes are being made in a
follow-up and usage of the added function is deferred until then.

Added tests to assert the new behavior.

Fixed a few tests around the stateless mode behavior where the Server ID
option used to be omitted, but are now required since we perform
validation.

Test: fx test dhcpv6-core-test dhcpv6-client-tests
Change-Id: I4c6c1ef7a16d53b3b66c3429cf5ac7acb91ecb65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/691943
Reviewed-by: Ghanan Gowripalan <ghanan@google.com>
Commit-Queue: Tony Gong <gongt@google.com>
Reviewed-by: Jake Fried <friedj@google.com>