[virtualization] Use llcpp in the gpu controller
To support both async events and sync method calls the gpu controller
was using 2 distinct connections to the gpu. This worked fine with the
c++ implementation, however the rust gpu device will not support
multiple connections in the manor.

Instead we can use the llcpp bindings that allows for this mix and match
of async events and sync method calls.

Bug: 102870
Test: fx test virtio_gpu_unittests
Change-Id: Iafa4e7e9571419522cc502e856d3ce78b5a6c763
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/699184
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>
Reviewed-by: David Hastings <dahastin@google.com>