[scenic][screen_capture2] implement manager
This CL adds the ScreenCapture manager, connected to the Frame
Scheduler, to call clients when OnCpuWorkDone is fired. It manages the
clients so that they are destroyed when either the manager closes or the
instance itself closes.

Unittests were added to ensure the manager correctly instantiates
clients, destroys clients, and calls clients when OnCpuWorkDone is
called. Functions were refactored from the screen capture unittests.

Fixed: 104355
Test: fx test screen_capture2_unittests
Change-Id: I5c65a2d52cd986736ef40f9df9ac02d1844cecc0
Commit-Queue: Oliver Lance <ollylance@google.com>
Reviewed-by: Emircan Uysaler <emircan@google.com>
Reviewed-by: Felipe Archondo <felipearchondo@google.com>
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/700083